final-pjt-back/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.7
django-allauth==0.50.0
django-cors-headers==3.12.0
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
oauthlib==3.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tmdbv3api==1.7.6
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
